Refactor mage tests into different source files
