+++ tools/mage/build_namespace.go
@@ -36,6 +36,9 @@ const swaggerGlob = "api/gateway/*/api.yml"
 // Build contains targets for compiling source code.
 type Build mg.Namespace
 
+// Other mage commands can use this to invoke build commands directly.
+var build = Build{}
+
 // API Generate API source files from GraphQL + Swagger
 func (b Build) API() {
 	mg.Deps(b.generateSwaggerClients, b.generateWebTypescript)
