+++ tools/mage/test_python.go
@@ -1,0 +1,93 @@
+package mage
+
+import (
+	"fmt"
+	"strings"
+
+	"github.com/magefile/mage/mg"
+	"github.com/magefile/mage/sh"
+)
+
+var (
+	pythonTests = []testTask{
+		{"python unit tests", testPythonUnit},
+		{"pylint", testPythonLint},
+		{"bandit (python security linting)", testPythonBandit},
+		{"mypy (python type checking)", testPythonMypy},
+	}
+	pylintSrcDisabled = []string{
+		"duplicate-code",
+		"fixme",
+		"missing-module-docstring",
+		"too-few-public-methods",
+	}
+	pylintTestsDisabled = append(pylintSrcDisabled,
+		"missing-class-docstring",
+		"missing-function-docstring",
+		"no-self-use",
+		"protected-access",
+	)
+)
+
+// Test and lint Python source code
+func (Test) Python() {
+	runTests(pythonTests)
+}
+
+func testPythonUnit() error {
+	args := []string{"-m", "unittest", "discover"}
+	if mg.Verbose() {
+		args = append(args, "--verbose")
+	} else {
+		args = append(args, "--quiet")
+	}
+
+	for _, target := range []string{"internal/core", "internal/compliance", "internal/log_analysis"} {
+		if err := runWithCapturedOutput(pythonLibPath("python3"), append(args, target)...); err != nil {
+			return fmt.Errorf("python unit tests failed: %v", err)
+		}
+	}
+
+	return nil
+}
+
+func testPythonLint() error {
+	// pylint - runs twice (once for src directories, once for test directories)
+	args := []string{"-j", "0", "--max-line-length", "140", "--score", "no"}
+	if mg.Verbose() {
+		args = append(args, "--verbose")
+	}
+
+	// pylint src
+	srcArgs := append(args, "--ignore", "tests", "--disable", strings.Join(pylintSrcDisabled, ","))
+	if err := sh.RunV(pythonLibPath("pylint"), append(srcArgs, pyTargets...)...); err != nil {
+		return fmt.Errorf("pylint source failed: %v", err)
+	}
+
+	// pylint tests
+	testArgs := append(args, "--ignore", "src", "--disable", strings.Join(pylintTestsDisabled, ","))
+	if err := sh.RunV(pythonLibPath("pylint"), append(testArgs, pyTargets...)...); err != nil {
+		return fmt.Errorf("pylint tests failed: %v", err)
+	}
+
+	return nil
+}
+
+func testPythonBandit() error {
+	args := []string{"--recursive"}
+	if mg.Verbose() {
+		args = append(args, "--verbose")
+	} else {
+		args = append(args, "--quiet")
+	}
+	return runWithCapturedOutput(pythonLibPath("bandit"), append(args, pyTargets...)...)
+}
+
+func testPythonMypy() error {
+	args := []string{"--cache-dir", "out/.mypy_cache", "--no-error-summary",
+		"--disallow-untyped-defs", "--ignore-missing-imports", "--warn-unused-ignores"}
+	if mg.Verbose() {
+		args = append(args, "--verbose")
+	}
+	return sh.RunV(pythonLibPath("mypy"), append(args, pyTargets...)...)
+}
